+++ config/webhook-server.php
@@ -8,9 +8,21 @@ return [
     'queue' => 'default',
 
     /*
+     *  The default queue connection that should be used to send webhook requests.
+     */
+    'connection' => null,
+
+    /*
      * The default http verb to use.
      */
     'http_verb' => 'post',
+
+    /*
+     * Proxies to use for request.
+     *
+     * See https://docs.guzzlephp.org/en/stable/request-options.html#proxy
+     */
+    'proxy' => null,
 
     /*
      * This class is responsible for calculating the signature that will be added to
@@ -32,7 +44,7 @@ return [
     ],
 
     /*
-     * If a call to a webhook takes longer that this amount of seconds
+     * If a call to a webhook takes longer this amount of seconds
      * the attempt will be considered failed.
      */
     'timeout_in_seconds' => 3,
@@ -48,10 +60,20 @@ return [
     'backoff_strategy' => \Spatie\WebhookServer\BackoffStrategy\ExponentialBackoffStrategy::class,
 
     /*
+     * This class is used to dispatch webhooks onto the queue.
+     */
+    'webhook_job' => \Spatie\WebhookServer\CallWebhookJob::class,
+
+    /*
      * By default we will verify that the ssl certificate of the destination
      * of the webhook is valid.
      */
     'verify_ssl' => true,
+
+    /*
+     * When set to true, an exception will be thrown when the last attempt fails
+     */
+    'throw_exception_on_failure' => false,
 
     /*
      * When using Laravel Horizon you can specify tags that should be used on the