+++ src/Events/DispatchingWebhookCallEvent.php
@@ -2,10 +2,7 @@
 
 namespace Spatie\WebhookServer\Events;
 
-use GuzzleHttp\Psr7\Response;
-use GuzzleHttp\TransferStats;
-
-abstract class WebhookCallEvent
+class DispatchingWebhookCallEvent
 {
     public function __construct(
         public string $httpVerb,
@@ -14,12 +11,7 @@ abstract class WebhookCallEvent
         public array $headers,
         public array $meta,
         public array $tags,
-        public int $attempt,
-        public ?Response $response,
-        public ?string $errorType,
-        public ?string $errorMessage,
         public string $uuid,
-        public ?TransferStats $transferStats
     ) {
     }
 }