+++ tests/TestClasses/TestClient.php
@@ -2,23 +2,27 @@
 
 namespace Spatie\WebhookServer\Tests\TestClasses;
 
+use GuzzleHttp\ClientInterface;
 use GuzzleHttp\Exception\ConnectException;
 use GuzzleHttp\Exception\RequestException;
+use GuzzleHttp\Promise\PromiseInterface;
 use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\Psr7\Response;
 use PHPUnit\Framework\Assert;
+use Psr\Http\Message\RequestInterface;
+use Psr\Http\Message\ResponseInterface;
 
-class TestClient
+class TestClient implements ClientInterface
 {
-    private array $requests = [];
+    protected array $requests = [];
 
-    private int $useResponseCode = 200;
+    protected int $useResponseCode = 200;
 
-    private bool $throwRequestException = false;
+    protected bool $throwRequestException = false;
 
-    private bool $throwConnectionException = false;
+    protected bool $throwConnectionException = false;
 
-    public function request(string $method, string $url, array $options)
+    public function request(string $method, $url = '', array $options = []): ResponseInterface
     {
         $this->requests[] = compact('method', 'url', 'options');
 
@@ -71,5 +75,25 @@ class TestClient
     public function throwConnectionException()
     {
         $this->throwConnectionException = true;
+    }
+
+    public function send(RequestInterface $request, array $options = []): ResponseInterface
+    {
+        throw new \BadMethodCallException('Not meant to be used yet.');
+    }
+
+    public function sendAsync(RequestInterface $request, array $options = []): PromiseInterface
+    {
+        throw new \BadMethodCallException('Not meant to be used yet.');
+    }
+
+    public function requestAsync(string $method, $uri, array $options = []): PromiseInterface
+    {
+        throw new \BadMethodCallException('Not meant to be used yet.');
+    }
+
+    public function getConfig(?string $option = null)
+    {
+        throw new \BadMethodCallException('Not meant to be used yet.');
     }
 }